custom-elements/instantiating-custom-elements/custom-element-type-local-name.html

<!DOCTYPE html>
<html>
<head>
<title>Custom element type is the local name of the custom element</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="The custom element type can be given to a custom element at the time of its instantation as the local name of the custom element">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#instantiating-custom-elements">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">
test(function() {
    var doc = newHTMLDocument();
    var GeneratedConstructor = doc.registerElement('x-a', {prototype: newHTMLElementPrototype('x-a')});
    var customElement = new GeneratedConstructor();
    assert_equals(customElement.localName, 'x-a',
        'Custom element type should be the local name of the custom element');
    assert_equals(customElement.constructor.prototype.typeName, 'x-a',
        'Custom element type should be the local name of the custom element');
}, 'Custom element type is the local name of the custom element. ' +
    'Custom element created via constructor');


test(function() {
    var doc = newHTMLDocument();
    doc.registerElement('x-b', {prototype: newHTMLElementPrototype('x-b')});
    doc.body.innerHTML = '<x-b id="x-b"></x-b>';
    var customElement = doc.querySelector('#x-b');
    assert_equals(customElement.constructor.prototype.typeName, 'x-b',
        'Custom element type should be the local name of the custom element');
}, 'Custom element type is the local name of the custom element.' +
    'Custom element created via innerHTML property');


testInIFrame('../resources/x-element.html', function(doc) {
    doc.registerElement('x-element', {prototype: newHTMLElementPrototype('x-element')});
    var xelement = doc.querySelector('#x-element');
    assert_equals(xelement.constructor.prototype.typeName, 'x-element',
        'Custom element type should be the local name of the custom element');
}, 'Custom element type is the local name of the custom element. ' +
    'Custom element is defined in loaded HTML document');
</script>
</body>
</html>
